add and implement spring data jpa service for pet. closes #41.

+++ pet-clinic-data/src/main/java/org/learning/spring/guruspringpetclinic/services/PetTypeService.java
@@ -2,5 +2,5 @@ package org.learning.spring.guruspringpetclinic.services;
 
 import org.learning.spring.guruspringpetclinic.model.PetType;
 
-public interface PetTypeService extends CrudService<PetType, Long> {
+public interface PetTypeService extends CrudService<PetType, Short> {
 }
